[refractor] break k8s client into smaller components
